requirements.txt
Django==1.11.24
djangorestframework==3.9.4
psycopg2-binary==2.8.3
nexmo
